Countries-Climate-indicators-Dataset-Visualization/
This data presents the temperature change with respect to a baseline for various countries and regions around the world, along with related environmental and climate change indicators. The data is collected by the Food and Agriculture Organization of the United Nations (FAO) and includes annual temperature change figures from the years 2013 to 2022. Each row represents a different country or region, and each column represents a specific year's temperature change data. It should be noted that for some countries, data was not available for a specific year, which is represented as 'NaN' in this dataset. This data provides a comprehensive overview of climate change in different regions and can be used to monitor changes over time.